Handle powers of/to scalar TensExpr objects
For example, it makes sense to divide by a tensor if its rank is zero (a
scalar), so it useful to allow such operations, which previously
resulted in an error.

*Discussion.*  The eventual goal is to allow scalar `TensExpr`s to be
wrapped by sympy objects that represent operations on scalar values like
Pow, sqrt, exp, etc.  We would conversely not want "scalar" sympy
operations to be able to apply themselves to `TensExpr`s that have
non-zero rank.  For exmaple, if `i` represents an *abstract* index of
the vector `x`, `exp(x(i) * x(-i))` makes sense but `exp(x(i))` does
not.  One might argue that you could define the latter to be an
componentwise/elementwise operation, but that involves fixing a specific
coordinate system to define the direction for each of the components.
In contrast, the idea of the abstract index notation implemented by this
module is to stay independent of a specific parameterization of space.